tests: move forgotten files
