+++ server/tests/End2End/Company/UI/Controller/EditCompanyTest.php
@@ -18,70 +18,33 @@ use End2End\Tests\DatabaseHelper;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 
-class PostCompanyTest extends AbstractControllerTest
+class EditCompanyTest extends AbstractControllerTest
 {
     /**
      * @throws \JsonException
      */
-    final public function testPostCompanySuccess(): void
+    final public function testPutCompanyFailWithBadUuid(): void
     {
         // Arrange
         $content = [
             'companyName' => 'Dev-Int Création',
-            'address' => '1, rue des ERP',
-            'zipCode' => '75000',
-            'town' => 'PARIS',
+            'address' => '2 rue des ERP',
+            'zipCode' => '56000',
+            'town' => 'VANNES',
             'country' => 'France',
             'phone' => '+33100000001',
             'facsimile' => '+33100000002',
             'email' => 'contact@developpement-interessant.com',
-            'contactName' => 'Laurent',
-            'cellphone' => '+33600000002',
+            'contactName' => 'Laurent Quétier',
+            'cellphone' => '+33100000002',
         ];
-        DatabaseHelper::loadFixtures([['group' => 'user']]);
-        $adminClient = $this->createAdminClient();
-        $adminClient->request(
-            Request::METHOD_POST,
-            '/api/companies/',
-            [],
-            [],
-            ['CONTENT_TYPE' => 'application/json'],
-            \json_encode($content, \JSON_THROW_ON_ERROR)
-        );
-
-        // Act
-        $response = $this->client->getResponse();
-
-        // Assert
-        self::assertSame(Response::HTTP_CREATED, $response->getStatusCode());
-        self::assertSame('Company create started!', $response->getContent());
-    }
-
-    /**
-     * @throws \JsonException
-     */
-    final public function testPostCompanyAlreadyExist(): void
-    {
-        // Arrange
         DatabaseHelper::loadFixtures([['group' => 'user'], ['group' => 'company']]);
-        $content = [
-            'companyName' => 'Dev-Int Création',
-            'address' => '1, rue des ERP',
-            'zipCode' => '75000',
-            'town' => 'PARIS',
-            'country' => 'France',
-            'phone' => '+33100000001',
-            'facsimile' => '+33100000002',
-            'email' => 'contact@developpement-interessant.com',
-            'contactName' => 'Laurent',
-            'cellphone' => '+33600000002',
-        ];
+        $adminClient = $this->createAdminClient();
 
         // Act
-        $adminClient = $this->createAdminClient();
         $adminClient->request(
-            Request::METHOD_POST,
-            '/api/companies/',
+            Request::METHOD_PUT,
+            '/api/companies/626adfca-fc5d-415c-9b7a-7541030bd147',
             [],
             [],
             ['CONTENT_TYPE' => 'application/json'],
@@ -92,7 +55,44 @@ class PostCompanyTest extends AbstractControllerTest
 
         // Assert
         self::assertSame(Response::HTTP_BAD_REQUEST, $response->getStatusCode());
-        self::assertSame(Request::METHOD_POST, $responseDecoded->method);
+        self::assertSame(Request::METHOD_PUT, $responseDecoded->method);
         self::assertSame('Company already exist', $responseDecoded->details);
+    }
+
+    /**
+     * @throws \JsonException
+     */
+    final public function testPutCompanySuccess(): void
+    {
+        // Arrange
+        $content = [
+            'companyName' => 'Dev-Int Création',
+            'address' => '2 rue des ERP',
+            'zipCode' => '56000',
+            'town' => 'VANNES',
+            'country' => 'France',
+            'phone' => '+33100000001',
+            'facsimile' => '+33100000002',
+            'email' => 'contact@developpement-interessant.com',
+            'contactName' => 'Laurent Quétier',
+            'cellphone' => '+33100000002',
+        ];
+        DatabaseHelper::loadFixtures([['group' => 'user'], ['group' => 'company']]);
+        $adminClient = $this->createAdminClient();
+        $adminClient->request(
+            Request::METHOD_PUT,
+            '/api/companies/a136c6fe-8f6e-45ed-91bc-586374791033',
+            [],
+            [],
+            ['CONTENT_TYPE' => 'application/json'],
+            \json_encode($content, \JSON_THROW_ON_ERROR)
+        );
+
+        // Act
+        $response = $this->client->getResponse();
+
+        // Assert
+        self::assertSame(Response::HTTP_FOUND, $response->getStatusCode());
+        self::assertSame('Company update started!', $response->getContent());
     }
 }
